requirements.txt
elevenlabs
python-dotenv==1.0.0
pyshorteners==1.0.1
requests==2.31.0
gunicorn==20.1.0
